+++ src/test/org/apache/commons/pool/impl/TestStackObjectPool.java
@@ -196,13 +196,13 @@ public class TestStackObjectPool extends TestBaseObjectPool {
         }
         // 3, 9, 15 fail passivation.  
         assertEquals(7,pool.getNumIdle());
-        assertEquals(new Integer(19), (Integer) pool.borrowObject());
-        assertEquals(new Integer(17), (Integer) pool.borrowObject());
-        assertEquals(new Integer(13), (Integer) pool.borrowObject());
-        assertEquals(new Integer(11), (Integer) pool.borrowObject());
-        assertEquals(new Integer(7), (Integer) pool.borrowObject());
-        assertEquals(new Integer(5), (Integer) pool.borrowObject());
-        assertEquals(new Integer(1), (Integer) pool.borrowObject());     
+        assertEquals(new Integer(19), pool.borrowObject());
+        assertEquals(new Integer(17), pool.borrowObject());
+        assertEquals(new Integer(13), pool.borrowObject());
+        assertEquals(new Integer(11), pool.borrowObject());
+        assertEquals(new Integer(7), pool.borrowObject());
+        assertEquals(new Integer(5), pool.borrowObject());
+        assertEquals(new Integer(1), pool.borrowObject());     
     }
     
     public void testBorrowReturnWithSometimesInvalidObjects() throws Exception {
